feat: 新增creta lib命令用于创建组件库

+++ packages/coco-cli/src/create.ts
@@ -5,14 +5,15 @@ import prompts from 'prompts';
 import path from 'path';
 import fse from 'fs-extra';
 
-async function createApp() {
+async function create(type: 'app' | 'lib') {
   let userCancelled = false;
   const response = await prompts(
     [
       {
         type: 'text',
         name: 'projectName',
-        message: '项目名称（新建文件夹，并且设置package.json的name）？',
+        message:
+          '项目名称（在当前目录下新建文件夹，且设置package.json的name）？',
       },
       {
         type: 'text',
@@ -49,30 +50,43 @@ async function createApp() {
     fse.removeSync(targetDir);
   }
 
-  const templateFolder = path.resolve(__dirname, '..', 'template');
+  const folder = type === 'app' ? 'app' : type === 'lib' ? 'lib' : '';
+  if (!folder) {
+    console.error('指定的模版目录不应该为空，', type);
+    return;
+  }
+  const tempFolderPath = path.resolve(__dirname, '..', 'template', folder);
+  if (!fse.pathExistsSync(tempFolderPath)) {
+    console.error('模版目录不存在', tempFolderPath, type);
+    return;
+  }
   // 复制非ejs文件
-  const noEjsFiles = glob.sync(`${templateFolder}/**/*`, {
+  const noEjsFiles = glob.sync(`${tempFolderPath}/**/*`, {
     ignore: '**/*.ejs',
     nodir: true,
   });
   for (const file of noEjsFiles) {
     const targetPath = path.join(
       targetDir,
-      path.relative(templateFolder, file)
+      path.relative(tempFolderPath, file)
     );
     fse.copySync(file, targetPath);
   }
 
   // 生成package.json
-  const packageJsonEjs = path.resolve(templateFolder, 'package.json.ejs');
+  const packageJsonEjs = path.resolve(tempFolderPath, 'package.json.ejs');
   const cliPackageJson = jsonfile.readFileSync(
     path.resolve(__dirname, '..', 'package.json')
   );
   // 目前coco-mvc和coco-cli的版本号保持一致
+  const version = cliPackageJson.version;
+  const transformerVersion =
+    cliPackageJson.devDependencies['@cocojs/typescript-transformer'];
   const renderedContent = await ejs.renderFile(packageJsonEjs, {
     name: projectName,
     author,
-    version: cliPackageJson.version,
+    version,
+    transformerVersion,
   });
   await fse.outputFileSync(
     path.join(targetDir, 'package.json'),
@@ -80,4 +94,5 @@ async function createApp() {
   );
 }
 
-export default createApp;
+export const createApp = () => create('app');
+export const createLib = () => create('lib');
